Made changes according to PR comments
